#44 fix test

+++ FactFactory/VersionedFactFactory/FactFactory.VersionedTests/VersionedFactRule/VersionedFactRuleTests.cs
@@ -125,7 +125,7 @@ namespace FactFactory.VersionedTests.VersionedFactRule
                 {
                     Assert.IsNotNull(fact.Version, "Version cannot be null");
                     if (fact.Version is Version2 version2)
-                        Assert.AreEqual((uint)2, version2.Value, "expected another version");
+                        Assert.AreEqual(2, version2.Value, "expected another version");
                     else
                         Assert.Fail("Version is not Version2");
                 });
